requirements.txt
rpaframework==15.0.0
google-api-python-client==2.23.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
ta-bitwarden-cli==0.7.5
pre-commit==2.17.0
ta-captcha-solver==0.2.0
supervisor==0.3.0
oauth2client==4.1.3
retry==0.9.2
Jinja2==3.1.2
supervisor==0.3.0